Fix #27 - Remove deprecations and warnings
This commit works to clear out deprecated code as well
as remove some warnings that get omitted as a result of
updates.

It resolves the following deprecations and warnings:

 * windows_package deprecated in windows cookbook
 * node.set deprecated to be removed in Chef 14
 * Warning about depecatd methods overriding providers from chefspec

In order to support these fixes, the cookbook now requires
at least chef 12.6 in order to use the new package resource.